+++ src/entrypoints/options/pages/video-subtitles/index.tsx
@@ -1,16 +1,20 @@
 import { i18n } from '#imports'
 import { PageLayout } from '../../components/page-layout'
-import { AutoStartSubtitlesToggle } from './auto-start-subtitles-toggle'
+import { ClearAiSegmentationCache } from './clear-ai-segmentation-cache'
+import { SubtitlesConfig } from './subtitles-config'
+import { SubtitlesRequestBatch } from './subtitles-request-batch'
+import { SubtitlesRequestRate } from './subtitles-request-rate'
 import { SubtitlesStyleSettings } from './subtitles-style-settings'
-import { VideoSubtitlesToggle } from './video-subtitles-toggle'
 
 export function VideoSubtitlesPage() {
   return (
     <PageLayout title={i18n.t('options.videoSubtitles.title')}>
       <div className="*:border-b [&>*:last-child]:border-b-0">
-        <VideoSubtitlesToggle />
-        <AutoStartSubtitlesToggle />
+        <SubtitlesConfig />
         <SubtitlesStyleSettings />
+        <SubtitlesRequestRate />
+        <SubtitlesRequestBatch />
+        <ClearAiSegmentationCache />
       </div>
     </PageLayout>
   )